LPFG-941: Merge branch 'LPFG-940-Create-DB-Tables'. Changed Long ids to ints. Changed LocalDateTime fields to Instant

+++ src/main/java/uk/gov/cslearning/record/service/BookingService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface BookingService {
     @Transactional(readOnly = true)
-    Optional<BookingDto> find(long bookingId);
+    Optional<BookingDto> find(int bookingId);
 
     @Transactional
     BookingDto save(BookingDto booking);
